+++ cmd/ssql-playground/main.go
@@ -208,6 +208,7 @@ func buildCommand() *cf.Command {
 	cmd = commands.RegisterDistinct(cmd)
 	cmd = commands.RegisterCount(cmd)
 	cmd = commands.RegisterDescribe(cmd)
+	cmd = commands.RegisterUnpivot(cmd)
 	cmd = commands.RegisterWhere(cmd)
 	cmd = commands.RegisterUpdate(cmd)
 	cmd = commands.RegisterCast(cmd)