+++ cmd/ssql/main.go
@@ -69,6 +69,7 @@ func buildRootCommand() *cf.Command {
 	cmd = commands.RegisterDistinct(cmd)
 	cmd = commands.RegisterCount(cmd)
 	cmd = commands.RegisterDescribe(cmd)
+	cmd = commands.RegisterUnpivot(cmd)
 	cmd = commands.RegisterWhere(cmd)
 	cmd = commands.RegisterUpdate(cmd)
 	cmd = commands.RegisterCast(cmd)